Refactored svg_writer
