+++ crates/vsvg/src/document/mod.rs
@@ -7,7 +7,7 @@ mod flattened_document;
 mod metadata;
 
 use crate::stats::LayerStats;
-use crate::svg_writer::document_to_svg_doc;
+use crate::writer::document_to_svg_doc;
 pub use document::Document;
 pub use flattened_document::FlattenedDocument;
 pub use metadata::DocumentMetadata;
