+++ crates/vsvg/src/svg/writer.rs
